requirements.txt
Flask==2.3.3
celery==5.3.4
redis==5.0.1
minio==7.1.17
pytesseract==0.3.10
Pillow==10.0.1
gunicorn==21.2.0
python-dotenv==1.0.0
werkzeug==2.3.7
flask-cors==4.0.0
pytest==7.4.2 